Final working prototype without required recent fixes
On 11.09.2020 server started to output json dictionary list with different structure, so I need to update my dictionary list processing algorithm in the next version or the program stops due to the currently chosen algorithm with the correct output to the log file. Ufter this update - it will be final version of the prototype. The following task when i complete final version - rewrite the program with different structure and make it work with multiple threads