Polishing working prototype (#2)
* Working on finishing the prototype with requested changes

This branch will lead to finishing the prototype according to the updated requirements and will have a lot of bad code(the whole project basically is bad code, but still - it works) and then in another branch i will take final functionality and stucture to make it work as it will in the end, but will refactor it from the start and rewrite a lot of code. This maybe will make this simple project worth a little more than hello world, but we will see that later

* Final working prototype without required recent fixes

On 11.09.2020 server started to output json dictionary list with different structure, so I need to update my dictionary list processing algorithm in the next version or the program stops due to the currently chosen algorithm with the correct output to the log file. Ufter this update - it will be final version of the prototype. The following task when i complete final version - rewrite the program with different structure and make it work with multiple threads

* Hopefully final version before refactoring

Final version - fixed wrong dictionaries list processing. Next task is to remake all code into something that looks better and works more reliable